+++ src/App.tsx
@@ -2,6 +2,8 @@ import { Navigate, Route, Routes } from 'react-router-dom'
 import { Onboarding } from './pages/Onboarding'
 import { SignIn } from './pages/SignIn'
 import { SignUp } from './pages/SignUp'
+import { ForgotPassword } from './pages/ForgotPassword'
+import { ResetPassword } from './pages/ResetPassword'
 import { Account } from './pages/Account'
 import { Dashboard } from './pages/Dashboard'
 import { Reserve } from './pages/Reserve'
@@ -14,6 +16,8 @@ function App() {
       <Route path="/" element={<Onboarding />} />
       <Route path="/connexion" element={<SignIn />} />
       <Route path="/inscription" element={<SignUp />} />
+      <Route path="/mot-de-passe-oublie" element={<ForgotPassword />} />
+      <Route path="/reinitialiser-mot-de-passe" element={<ResetPassword />} />
       <Route
         path="/tableau-de-bord"
         element={